+++ R/bundlePackage.R
@@ -3,13 +3,15 @@ bundlePackages <- function(
   extraPackages = character(),
   quiet = FALSE,
   verbose = FALSE,
+  ignoreLockfile = FALSE,
   error_call = caller_env()
 ) {
   deps <- computePackageDependencies(
     bundleDir,
     extraPackages,
     quiet = quiet,
-    verbose = verbose
+    verbose = verbose,
+    ignoreLockfile = ignoreLockfile
   )
   if (nrow(deps) == 0) {
     return(list())
@@ -45,7 +47,8 @@ computePackageDependencies <- function(
   bundleDir,
   extraPackages = character(),
   quiet = FALSE,
-  verbose = FALSE
+  verbose = FALSE,
+  ignoreLockfile = FALSE
 ) {
   if (usePackrat()) {
     taskStart(quiet, "Capturing R dependencies with packrat")
@@ -58,12 +61,15 @@ computePackageDependencies <- function(
       extraPackages,
       verbose = verbose
     )
-  } else if (!is.null(resolveRenvLockFile(bundleDir))) {
+  } else if (!ignoreLockfile && !is.null(resolveRenvLockFile(bundleDir))) {
     lockfile <- resolveRenvLockFile(bundleDir)
     # This ignores extraPackages; if you're using a lockfile it's your
     # responsibility to install any other packages you need
     taskStart(quiet, "Capturing R dependencies from renv.lock")
-    deps <- parseRenvDependencies(lockfile, bundleDir)
+    deps <- parseRenvDependencies(
+      lockfile,
+      bundleDir
+    )
     # Once we've captured the deps, we can remove the renv directory
     # from the bundle (retaining the renv.lock).
     removeRenv(bundleDir, lockfile = FALSE)